requirements.txt
# DocChat — Python Dependencies
# Run: pip install -r requirements.txt

# Core framework
streamlit==1.35.0

# LangChain ecosystem
langchain==0.2.1
langchain-openai==0.1.7
langchain-huggingface==0.0.3
sentence-transformers==3.0.0
langchain-community==0.2.1

# LLM providers
openai==1.30.1

# PDF parsing
pypdf==4.2.0

# DOCX support
python-docx==1.1.0

# Vector similarity search
faiss-cpu==1.8.0

# Tokenisation
tiktoken==0.7.0

# Environment variable loading
python-dotenv==1.0.1